Move DateSerializer to Storage class
